Something

+++ tools/__init__.py
@@ -1,5 +1,6 @@
 import contextvars
 import json
+import os
 import shutil
 
 from config.database import IN_MEMORY_DB, TINYDB_PARAMS
@@ -18,27 +19,14 @@ class _BytesJSONEncoder(json.JSONEncoder):
         return super().default(obj)
 
 
-CLUSTER_CONTEXT = contextvars.ContextVar("cluster_context", default=None)
-
-CLUSTER_TASKS = [
-    "users_create_user",
-    "users_create_credential",
-    "users_user_patch",
-    "users_user_patch_profile",
-    "users_user_patch_credential",
-    "users_user_delete",
-    "users_user_delete_credential",
-    "objects_object_create",
-    "objects_object_patch",
-    "objects_object_delete",
-]
+CONTEXT_TRANSACTION = contextvars.ContextVar("context_transaction", default=None)
 
 
 def evaluate_db_params(kwargs):
     db_params = copy(TINYDB_PARAMS)
 
-    if CLUSTER_CONTEXT.get():
-        transaction_file = f"database/main.{CLUSTER_CONTEXT.get()}"
+    if CONTEXT_TRANSACTION.get():
+        transaction_file = f"database/main.{CONTEXT_TRANSACTION.get()}"
     elif kwargs.get("transaction"):
         transaction_file = "database/main.{transaction}".format(
             transaction=float(kwargs.get("transaction"))
@@ -48,7 +36,8 @@ def evaluate_db_params(kwargs):
 
     if transaction_file != TINYDB_PARAMS["filename"]:
         assert is_path_within_cwd(transaction_file)
-        shutil.copy(TINYDB_PARAMS["filename"], transaction_file)
+        if not os.path.exists(transaction_file):
+            shutil.copy(TINYDB_PARAMS["filename"], transaction_file)
 
     db_params["filename"] = transaction_file
 
@@ -60,23 +49,21 @@ def cluster_task(task_name, enforce_uuid: bool = False):
         @wraps(func)
         async def wrapper(*args, **kwargs):
             result = await func(*args, **kwargs)
-            if not CLUSTER_CONTEXT.get():
+            if not CONTEXT_TRANSACTION.get():
                 return result
 
             self = args[0] if args else None
 
-            try:
-                if enforce_uuid:
-                    self.init_kwargs.update({"_enforce_uuid": result})
+            if enforce_uuid:
+                self.init_kwargs.update({"_enforce_uuid": result})
 
-                task_request = "TASK {task_name} [{init_kwargs}, {task_kwargs}]".format(
-                    task_name=task_name,
-                    init_kwargs=json.dumps(self.init_kwargs, cls=_BytesJSONEncoder),
-                    task_kwargs=json.dumps(kwargs, cls=_BytesJSONEncoder),
-                )
-                TaskModel.parse_raw_task(task_request)
-            except (ValidationError, ValueError) as e:
-                print("Task validation error:", e)
+            task_request = "TASK {task_name} [{init_kwargs}, {task_kwargs}]".format(
+                task_name=task_name,
+                init_kwargs=json.dumps(self.init_kwargs, cls=_BytesJSONEncoder),
+                task_kwargs=json.dumps(kwargs, cls=_BytesJSONEncoder),
+            )
+
+            TaskModel.parse_raw_task(task_request)
 
             if task_request not in IN_MEMORY_DB["tasks"]:
                 IN_MEMORY_DB["tasks"].append(task_request)
